==> index/03-data.Rmd <==
# Data {#data}

```{r dat, include=FALSE}

library(knitr)     # for knitting
library(kableExtra)
library(dplyr)     # used throughout
library(ggplot2)
library(stringr)
library(reshape2)
library(R2jags)    # for MCMC
library(png)       # for graphics
library(grid)
library(gridExtra)

knitr::opts_chunk$set(warning = FALSE, message=FALSE, echo=FALSE)
#, fig.pos='H')

load("../rdatafiles/Xtot.RData")


id1 <- 887665 
id2 <- 842301 
id3 <- 603106 
id4 <- 842296 
playerseasons <- matrix(
  c(id1,   2016,
    id2,   2015,
    id3,   2014,
    id4,   2016
  ),ncol=2,byrow = TRUE
)
colnames(playerseasons) <- c("globalplayerid" ,"season")



tabsamp <- Xtot %>%
  '['(700:704,) %>%
  mutate(gameid = as.character(gameid),
         time   = as.character(time),
         r      = round(r, 4),
         theta  = round(theta, 4)) %>%
  select(season, gameid, time, globalplayerid, r, theta, home, result)

tabsum <- matrix(nrow=4, ncol=ncol(tabsamp))
colnames(tabsum) <- colnames(tabsamp)
rownames(tabsum) <- c("Name","Type", "Values", "Extra Details")
tabsum[,1] <- c("season", "categorical", "{2014, ..., 2017}", "")
tabsum[,2] <- c("gameid","categorical", "NA", "94 unique values")
tabsum[,3] <- c("time","continuous", "NA", "13-digit timestamp in milliseconds")
tabsum[,4] <- c("globalplayerid","categorical", "NA", "31 unique values")
tabsum[,5] <- c("r","continuous", "[0, $\\infty$)", "Distance of shot from basket (feet)")
tabsum[,6] <- c("theta","continuous ", "[-$\\pi$, $\\pi$]", "Angle of shot (radians)")
tabsum[,7] <- c("home","categorical", "{0,1}", "1 if shot occured in a home game")
tabsum[,8] <- c("result","categorical", "{0,1}", "1 if shot was made (response)")
colnames(tabsum) <- NULL
tabsum <- t(tabsum)

ggblank <- ggplot(as.data.frame(1)) + 
  geom_blank() + 
  theme_bw() + 
  theme(panel.border =element_blank())
```


## Description of Dataset

The data for this analysis comes from SportVU, a player-tracking system from STATS, LLC. that provides precise coordinates for all ten players and the ball at a rate of 25 times per second. The Duke University Men's Basketball team permitted us to use their SportVU data from the 2014 to 2017 basketball seasons for this project. Since the ability to record this data depends on specialized tracking cameras, Duke does not have this data for every game they play---only home games, and a few road games in arenas that had the technology installed. Therefore, there is a substantial amount of missing data between games. More specifically, between the 2014 and 2017 seasons, the Duke Men's Basketball team played 147 games; this dataset contains `r n_distinct(Xtot$gameid)` games, with `r n_distinct(Xtot$gameid[which(Xtot$home==1)])` at Duke and `r n_distinct(Xtot$gameid[which(Xtot$home==0)])` at other arenas.

For our analysis, we use the following files for each game:

* Final Sequence Play-by-Play Optical:

This dataset comes in an a semi-structured Extensible Markup Language (XML) file, where there is a unique element for each "event" (an event is a basketball action such as a dribble, pass, shot, foul, etc.). Each event element has attributes describing the type of event, the time of the event, and the player who completed the action. We use these files to uncover when a shot is attempted in a game, who attempted the shot, and the result of the shot attempt.
    
<!--
* Box Score Optical:

We use this dataset to match the names and IDs of players who are in the game. This is also an XML file, with elements corresponding to individual players. These elements contain attributes describing information about the player (e.g. team name, jersey number) and various statistics for the game (e.g. points, assists, distance run).
-->

* Final Sequence Optical:

  These XML files contain the locations of all ten players and the ball during precise time intervals within the game (25 times per second). Each time unit has a unique element, and these elements have attributes describing the locations. We merge this with the Final Sequence Play-by-Play Optical data on the time attribute to obtain the shooter's location at the moment of a shot attempt.

## Data Cleaning

  Steps taken to clean the merged shooter IDs with shot locations include translating the locations to a half-court setting (the teams switch sides of the court halfway through every game, which means that we have to flip the coordinates across the middle of the court for about half of the shots in every game), converting the x-y coordinates to polar coordinates (in the units of feet and radians), and adding an indicator for home games. We only use the shots that Duke players attempt, because there is an inadequate amount of data for players on other teams---no opposing players appear in more than 5 games. The final dataset contains 5,467 observations from `r n_distinct(Xtot$globalplayerid)` shooters over `r n_distinct(Xtot$gameid)` games. A summary of the cleaned dataset is in Table \@ref(tab:summarytable):

```{r summarytable, results="asis"}
kable(tabsum, 
      format="pandoc",
      caption = "Summary of Dataset",
      caption.short = "Data Summary",
      longtable = TRUE,
      booktabs = FALSE)
```

A small subset of the cleaned data is displayed below in Table \@ref(tab:sampletable):

```{r sampletable, results="asis"}
kable(tabsamp,
      caption = "Sample of Dataset",
      caption.short = "Data Sample",
      longtable = TRUE,
      booktabs = TRUE) %>%
  kable_styling(latex_options = "scale_down") %>%
  kable_styling(font_size=11)
```




\pagebreak

Figure \@ref(fig:shotplot) shows the locations of all the shots in the dataset, translating the locations to one half of the court, and excluding heaves from beyond half court. The variable $\theta$ has a range of $2\pi$ radians, but this plot shows that most of the attempts occur within the interval (-$\frac{\pi}{2}$, $\frac{\pi}{2}$). This figure also shows the bimodal distribution of shot distance over all players.

```{r shotplot, fig.cap = "Locations and Results of All Shots"}
img.path <- "./figure/ncaa_bball_court2.png"
img <- readPNG(img.path)

xrange <- c(-25,25)
yrange <- c(0,94/2) - 4
Xtotrange <- Xtot %>%
  filter(yt > yrange[1] & yt < yrange[2] & xt > xrange[1] & xt < xrange[2])

plt <- ggplot(Xtotrange, aes(x=xt, y=yt, color=as.factor(result))) + 
  annotation_custom(rasterGrob(img,
                               x=0.50, y=0.50,
                               width = .92, height = .92),
                    -Inf, Inf, -Inf, Inf) + 
  scale_color_manual(values=c("red", "blue"), labels=c("misses", "makes")) +
  geom_point(alpha=0.2) +
  geom_vline(xintercept = 0) +
  geom_hline(yintercept = 0) +
  geom_label(data = data.frame(c(NA, NA, NA)),
           x=c(-13, 3, 13), 
           y=c(-1, 40, -1), 
           parse=TRUE,
           label=c(as.character(expression(theta == frac(pi,2))),
                   as.character(expression(theta == 0)),
                   as.character(expression(theta == -frac(pi,2)))),
           size=3,
           color="black", 
           fontface="bold") +
  labs(title = "Distribution of Shot Locations", y="", x="", color = "Result") +
  theme_bw() + 
  theme(panel.grid = element_blank(), panel.border = element_blank(), axis.text = element_blank(), axis.ticks = element_blank())

plt
```


## Exploratory Data Analysis

The exploratory data analysis plots in Figure \@ref(fig:smoothplot) examine how consistent the probability of a made shot is, using a loess smooth curve on the binary outcomes. We present these smoothed plots for four high-usage basketball players at Duke University between the 2014 and 2017 seasons. Each plot represents a single player's ordered shooting outcomes for a single season. These plots do not account for the amount of time in between shots, but simply shot order and outcome. 



```{r smoothplot0, fig.height=5, fig.width=5, fig.align="center"}

plotshottime <- function(playerid, season, main){
  Xtot_sub <- Xtot[Xtot$globalplayerid %in% playerid & Xtot$season %in% season,]
  par(las=1)
  if(nrow(Xtot_sub) < 10){
    # ggplot(data=NULL) + labs(title = "Not Enough Data") + theme_bw()
    plot(0,0,type="n", yaxt="n",xaxt="n", ylab="",xlab="")
    text(0,0,label="Not Enough Data", cex=2)
  }else{
    Y <- Xtot_sub$result #zoo::rollmean((Xtot_sub$result), 4)
    X <- 1:length(Y)
    scatter.smooth(X,Y,span=20/length(Y), main = main,
                   yaxt="n",ylab="Result",xlab="Order")
    axis(side=2,at=c(0,1),labels=c("Miss","Make"))
  }
}

maxseason <- Xtot %>% group_by(globalplayerid, season) %>% summarize(num=n()) %>% group_by(globalplayerid) %>% mutate(m=max(num)) %>% filter(num==m) %>% as.data.frame()

par(mfrow = c(2,2))
for(i in 1:nrow(playerseasons)){
  r <- playerseasons[i,]
  plotshottime(r[[1]],r[[2]], paste0("Player ",i))
}
```

```{r smoothplot, fig.cap = "Moving Average of Shot Success Rate", fig.height=3.5, fig.width=5, fig.align="center"}
plotshottime(Xtot$globalplayerid, Xtot$season, "Team Average")
```

\pagebreak

We can see that the players vary in the consistency of their made shots, since they all contain spikes and trends. For example, Player 3 initially has a very high success rate, which quickly falls to the middle after about 30 shot attempts, and the Player 2 has a noticeable upward trend in shot success beginning around shot number 150.

We investigate the shooting outcomes using Bayesian models, and present the results in the next chapter.
