==> index/05-disc.Rmd <==
# Discussion {#disc}

<!--
Our results so far suggest that some insights can be taken by predicting basketball shooting outcomes, but none of our results are more insightful than common intuition (i.e. the most frequent shooters are usually above average shooters, or the probability of making a shot decreases as distance from the basket increases). None of the models that we have built predict the out-of-sample data well. We could refine our models by including more predictors such as a proxy for fatigue (using information about total minutes played, or consecutive minutes played without a timeout), or shot difficulty (using information about the nearest defender). -->

## Evaluation of Models

To evaluate these models, we use 5-fold cross-validation. In each train-test split, we evaluate the models' out-of-sample classification rates (using a cutoff probability of 0.5), Brier scores (mean squared error), and log-likelihoods. The predictions and fitted values are obtained using MCMC averages; to calculate the probability for an individual shot, we calculate a response for each of the 9,500 posterior simulations, then take the average of those responses. This process used up to 20 simultaneous RStudio Pro servers provided by the Duke University Statistical Science Department. The results are plotted below in Figure \@ref(fig:evalplot):

```{r, fig.height=3.5}
ggblank
```

\pagebreak

```{r predfxns, cache=TRUE}
# posterior mean   --> minimize sq err
# posterior median --> minimize abs err

# we want S predictions per shot. an N x S matrix. then we take MCMC average.
pred_glm <- function(Xpred = NA, MCMC = NA){
  
  Xmat <- Xpred %>% 
    mutate(logr = log(r),
           away = 1-home) %>%
    arrange(time) %>%
    select(away, home, logr, theta) %>% 
    as.matrix()

  # the baseline 1 intercept = "away"
  # we should only multiply by that one if home = 0
  #params.glm <- t(MCMC)
  A <- Xmat
  B <- t(MCMC)
  pred.glm <-  arm::invlogit(A %*% B) %>% apply(., 1, median) #MCMC median
  
  return(as.matrix(pred.glm))
}

pred_me <- function(Xpred = NA, MCMC = NA){

  Xmat <- Xpred %>% 
    mutate(logr = log(r),
           away = 1-home) %>%
    arrange(time) %>%
    select(away, home, logr, theta, globalplayerid) %>% 
    as.matrix()

  pred.me <- array(NA, c(nrow(Xmat), 1))
  for(r in 1:nrow(Xmat)){
    gpid <- Xmat[r,"globalplayerid"]
    Asub <- Xmat[r,1:4] # 1 x p
    playerinds <- grep(gpid, colnames(MCMC))
    Bsub <- t(MCMC[,playerinds]) # p x S
    if(length(playerinds) == 0){
      Bsub <- t(MCMC[,which(!grepl("\\]", colnames(MCMC)))])
    }
    pred.me[r,] <- median(arm::invlogit(Asub %*% Bsub)) #MCMC avg
  }

  return(pred.me)
}

pred_disc <- function(Xpred = NA, MCMC.LIST=NA){
  
  pred.disc.large <- lapply(X=MCMC.LIST, FUN=pred_me, Xpred=Xpred) %>% 
    as.data.frame() %>%
    as.matrix()
  
  rownames(pred.disc.large) <- Xpred$gameid   #N
  colnames(pred.disc.large) <- gamemap$gameid #G, from the list
  
  pred.disc <- array(NA, c(nrow(Xpred), 1))

  for(i in 1:nrow(Xpred)){
    keepcol <- colnames(pred.disc.large) == Xpred[i,"gameid"]
    pred.disc[i,] <- pred.disc.large[i, keepcol]
  }

  return(pred.disc)
  
}


if(!load_chains){
  
  pred.team     <- pred_glm(Xtest, glmtot)
  pred.player   <- pred_me(Xtest, player.mcmc)
  eval(parse(
    text=paste0("pred.game.",deltas_str," <- pred_disc(Xtest, game.mcmc.list.", deltas_str, ");\nprint(",deltas_str,");")
  ))
  savepreds <- ls() %>% grep(pattern="pred.",fixed=TRUE,value=TRUE)
  temp <- lapply(as.list(savepreds), function(x){
    save(list=x, 
         file=paste0(rdatafiles,gsub(patt=".",rep="",x=x,fixed=TRUE),".RData"))
    })
  
  fitted.team     <- pred_glm(Xtrain, glmtot)
  fitted.player   <- pred_me(Xtrain, player.mcmc)
  eval(parse(
    text=paste0("fitted.game.", deltas_str," <- pred_disc(Xtrain, game.mcmc.list.", deltas_str, ");\nprint(",deltas_str,");")
  ))
  savefitteds <- ls() %>% grep(pattern="fitted.",fixed=TRUE,value=TRUE)
  temp <- lapply(as.list(savefitteds), function(x){
    save(list=x, 
         file=paste0(rdatafiles,gsub(patt=".",rep="",x=x,fixed=TRUE),".RData"))
  })
  

}else{
  
  prefixes <- c("pred", "fitted")
  suffixes <- paste0(c("team", "player", paste0("game",deltas_str)),".RData")
  loadfiles <- merge(prefixes, suffixes) %>% apply(1, paste0, collapse="") %>% paste0(rdatafiles, .)
  temp <- lapply(as.list(loadfiles), load, envir=globalenv())

}

```

```{r evalfxns, cache=TRUE}

#confusion matrix, classification rate, and brier score
classify <- function(observed = NA, predicted = NA, cutoff = 0.5, gameids = NA){

  eval.list <- list()
  
  if(is.na(gameids[1])){
    gameids <- 1
  }
  
  for(i in 1:n_distinct(gameids)){
    g <- unique(gameids)[i]
    pred <- predicted[gameids == g]
    obs <- observed[gameids == g]
    
    tab <- array(NA, c(2,2))
    rownames(tab) <- c("predicted miss", "predicted make")
    colnames(tab) <- c("observed miss", "observed make")
    pred01 <- ifelse(pred < cutoff, 0, 1)
    
    tab[1,1] <- sum(obs==0 & pred01==0)
    tab[1,2] <- sum(obs==1 & pred01==0)
    tab[2,1] <- sum(obs==0 & pred01==1)
    tab[2,2] <- sum(obs==1 & pred01==1)
    
    rate <- (tab[1,1]+tab[2,2])/length(obs)
    
    eval.list[[as.character(g)]] <-  rate
  }
    
  if(gameids == 1){
    return(rate)
  }else{
    return(eval.list)
  }
}

meansquare <- function(observed = NA, predicted = NA, gameids = NA){

  eval.list <- list()
  
  if(is.na(gameids[1])){
    gameids <- 1
  }
  
  for(i in 1:n_distinct(gameids)){
    g <- unique(gameids)[i]
    pred <- predicted[gameids == g]
    obs <- observed[gameids == g]
    
    MSE <- mean((obs - pred)^2)
    
    eval.list[[as.character(g)]] <-  MSE
  }
    
  if(gameids == 1){
    return(MSE)
  }else{
    return(eval.list)
  }
}

loglikelify <- function(observed = NA, predicted = NA, gameids = NA){

  eval.list <- list()
  
  if(is.na(gameids[1])){
    gameids <- 1
  }
  
  for(i in 1:n_distinct(gameids)){
    g <- unique(gameids)[i]
    pred <- predicted[gameids == g]
    obs <- observed[gameids == g]
    
    #deviance
    loglik <- sum(
      obs*log(pred) + (1-obs)*log((1-pred))
    )
    eval.list[[as.character(g)]] <-  loglik
  }
  
  if(gameids == 1){
    return(loglik)
  }else{
    return(eval.list)
  }
}

#data for calibration plots
calibrate <- function(observed = NA, predicted = NA, nbins = 20, gameids = NA){

  eval.list <- list()
  
  if(is.na(gameids[1])){
    gameids <- 1
  }
  
  for(i in 1:n_distinct(gameids)){
    g <- unique(gameids)[i]
    pred <- predicted[gameids == g]
    obs <- observed[gameids == g]

    if(n_distinct(pred) < nbins){
      calib <- NA
    }else{
      
      predquants <- c(0, quantile(pred, (1:nbins)/nbins))
      predmidpts <- ((predquants + c(0,predquants[-length(predquants)]))/2) %>% '['(-1)
    
      pred1N <- sapply(pred, findInterval, vec=predquants, all.inside=TRUE, rightmost.closed=TRUE, left.open=TRUE)
  
      calibtab <- table(pred1N, obs)
      rownames(calibtab) <- predmidpts
      calib <- calibtab %>% 
        as.data.frame() %>%
        dcast(pred1N ~ obs, value.var = "Freq") %>%
        mutate(p  = `1`/(`0`+`1`),
               pl = p - 2*sqrt(p*(1-p)/(`0` + `1`)),
               pu = p + 2*sqrt(p*(1-p)/(`0` + `1`)),
               binl = predquants[-(nrow(.)+1)],
               binu = predquants[-1])
      colnames(calib) <- c("bin", "obs0", "obs1", "p", "pl", "pu", "binl", "binu")

    }
    
    eval.list[[as.character(g)]] <- calib
  }
  
  if(gameids == 1){
    return(calib)
  }else{
    return(eval.list)
  }
}

plot_calibration <- function(dat, title="Calibration Plot"){
  
  ggplot(data = dat, aes(x=as.numeric(as.character(bin)), y=p, group=0)) + 
    geom_errorbar(aes(ymin=pl, ymax=pu), width=0) + 
    geom_errorbarh(aes(xmin=binl, xmax=binu), height=0) + 
    geom_abline(slope=1, color="red") +
    coord_cartesian(xlim=c(0,1), ylim=c(0,1)) + 
    theme(panel.grid = element_blank()
          #,panel.border = element_blank(), axis.text = element_blank(), axis.ticks =   element_blank()
          ) + 
    theme_bw() + 
    labs(x="Bins",y="Empirical Probability" , title=title)
}

if(!load_chains){
  
  class.team  <- classify(Xtest$result, pred.team)
  class.player <- classify(Xtest$result, pred.player)
  eval(parse(
    text=paste0("class.games <- c(",
           paste0("classify(Xtest$result,pred.game.",deltas_str,collapse="),"),"))"
         )
  ))

  mse.team  <- meansquare(Xtest$result, pred.team)
  mse.player <- meansquare(Xtest$result, pred.player)
  eval(parse(
    text=paste0("mse.games <- c(",
           paste0("meansquare(Xtest$result, pred.game.",deltas_str,collapse="),"),"))"
          )
    ))
  
  lik.team  <- loglikelify(Xtrain$result, fitted.team)
  lik.player <- loglikelify(Xtrain$result, fitted.player)
  eval(parse(
    text=paste0("lik.games <- c(",
           paste0("loglikelify(Xtrain$result, fitted.game.",deltas_str,collapse="),"),"))"
          )
    ))
  eval.df <- data.frame(
    class = c(class.team, class.player, class.games),
    mse = c(mse.team, mse.player, mse.games),
    lik = c(lik.team, lik.player, lik.games),
    type = c("GLM", "Mixed Effects",rep("Discount Likelihood", length(deltas))),
    delta = c(NA, NA, deltas_str),
    k = k0
  )

  save(eval.df, file=paste0(rdatafiles, "evaldf.RData"))
  
}else{
  
  tempenv <- new.env()
  eval.df <- NULL
  
  for(i in 1:k){
    
    load(file = paste0(get_rdatafiles(i), "evaldf.RData"), tempenv)
    eval.df <- rbind(eval.df, tempenv$eval.df)
    
  }
    
  eval.df$k <- as.factor(eval.df$k)
  eval.df$delta <- factor(eval.df$delta,sort(as.character(levels(eval.df$delta)))) #making sure that the deltas are in order

}

df_delta <- eval.df %>% filter(!is.na(delta))
df <- eval.df %>% filter(is.na(delta))
df_delta$delta <- df_delta$delta %>%
  as.character() %>%
  as.numeric() %>%
  '/'(1000) %>%
  as.character() %>%
  str_pad(width=5, side="right", pad="0") %>%
  as.factor()
  
g1 <- ggplot(data=df_delta, aes(x=delta,y=class, col=k)) +
   geom_point() + 
   geom_hline(data=df, aes(yintercept=class, linetype=type, col=k)) +
   theme_bw() + 
   theme(axis.text.x = element_text(angle = 45, hjust=1)) +
   labs(y="Classification Rate \n (0.5 cutoff)", x=expression(delta), linetype = "")


g2 <- ggplot(data=df_delta, aes(x=delta,y=lik, col=k)) +
  geom_point() +
  geom_hline(data=df, aes(yintercept=lik, linetype=type, col=k)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(y="Log-Likelihood \n", x=expression(delta), linetype = "")

g3 <- ggplot(data=df_delta, aes(x=delta,y=mse, col=k)) +
  geom_point() +
  geom_hline(data=df, aes(yintercept=mse, linetype=type, col=k)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(y="Brier Score (MSE)", x=expression(delta), linetype = "")
```

```{r evalplot, fig.cap="Model Evaluation", fig.height=4.75, fig.align="center"}
grid.arrange(g1, g2, g3, nrow=2)
```


From Figure \@ref(fig:evalplot), we can observe that all of the models have different strengths. The discounted likelihood model with the smallest value of $\delta$ consistently has the highest likelihood. However, it does not test as well as the other models in areas of out-of-sample classification rate and Brier score. This suggests that models with smaller values of $\delta$, where the likelihood of an observed shot is more heavily influenced by shots closer to it, may overfit the model to the training data. The generalized linear models perform best in Brier score, but worst in log-likelihood. The hierarchical models are about the same as the GLMs, but they have a better log-likelihood performance. A model that balances the trade-off between predictive accuracy and likelihood is a discounted likelihood model with $\delta$ = 0.850.

In addition, we can see that the overall variation in model performance is small. For example, most of the out-of-sample classification rates fall between 0.58 and 0.62. This is within the 95% confidence interval for a random binomial proportion of 0.6 using a sample size of `r n <- k*(length(deltas) + 2); n` (because there are `r length(deltas)+2` different models and `r k` train-test splits for each model), which is `r err <- sqrt(0.6*(1-0.6)/n); paste0("(", round(0.6-err, 4), ", " , round(0.6+err, 4), ")")`. Therefore, the evidence that the models without discounting predict better than the ones with discounting is not particularly strong.

For the discounted likelihood model with $\delta$ = 0.850, we build calibration plots to assess how well the estimated probabilities fit the actual proportions. To make these plots, we divide the predicted probabilities into 20 equally-sized bins, then plot these bins on the x-axis with the proportions of the actual outcomes within the bins on the y-axis. The horizontal bars represent the bin width, and the vertical bars represent a 95% confidence interval of the proportions. The red line of slope 1 represents equality between the bin medians and the empirical probabilities within the bins. In Figure \@ref(fig:caliplot), we present these plots for a full training set and a testing set.


```{r caliplot, fig.cap="Calibration Plots for Discounted Likelihood Model, $\\delta$ = 0.850"}

randtrain <- sample(Xtrain$gameid, 1)
randtest  <- sample(Xtest$gameid, 1)
randtrainrows <- Xtrain$gameid == randtrain
randtestrows  <- Xtest$gameid == randtest

calitrain <- Xtrain$result %>% 
  calibrate(observed=., predicted=fitted.game.850) %>% 
  plot_calibration(., "Training Set")

calitest <- Xtest$result %>% 
  calibrate(observed=., predicted=pred.game.850) %>% 
  plot_calibration(., "Testing Set")

# calitrain1 <- Xtrain$result[randtrainrows] %>% 
#   calibrate(observed=., predicted=fitted.game.850[randtrainrows], nbins=5) %>% 
#   plot_calibration(., "Training Game")
# 
# calitest1 <- Xtest$result[randtestrows] %>% 
#   calibrate(observed=., predicted=pred.game.850[randtestrows]) %>% 
#   plot_calibration(., "Testing Game")

grid.arrange(calitrain, calitest, ncol=2)

```

\pagebreak

We can see that the confidence intervals on the training set all cross the line of slope 1, which shows that the model output reliably fits the probabilities. In the testing set, however, the predictions only cross this line between about 0.3 and 0.75. In addition, the widths of the bins on the edges show that the model is not likely to predict values close to 0 or 1.

## Results from Model with $\delta$ = 0.850

To illustrate results from the discounted likelihood model with $\delta$ = 0.850, we replicate the plots in Figures \@ref(fig:discplot750) and \@ref(fig:discplot999). The results are shown in Figure \@ref(fig:discplot850).

```{r discplot850, fig.cap="Parameters for Two Players and Population over Time, $\\delta$ = 0.850", out.width="90%", fig.align="right"}

game.params.850_sub <- game.params.850 %>% 
  mutate(season = get_season(gameid)) %>% 
  filter(season == 2016)
IDS <- c(0, id1, id4) 
z_IDS <- c("Team", "Player 1", "Player 4")

discplot850H <- plot_time_effect(game.params.850_sub, ids=IDS, parameters="intH", main="Intercept (Home)", zlab = z_IDS, yrange = c(-2, 4))
discplot850A <- plot_time_effect(game.params.850_sub, ids=IDS, parameters="intA", main="Intercept (Away)", zlab = z_IDS, yrange = c(-2, 4))
discplot850r <- plot_time_effect(game.params.850_sub, ids=IDS, parameters="r", main="Distance", zlab = z_IDS, yrange = c(-1, 0.1))
discplot850theta <- plot_time_effect(game.params.850_sub, ids=IDS, parameters="theta", main="Angle", zlab = z_IDS, yrange = c(-1.5, 0.75))

grid.arrange(discplot850H,discplot850A,discplot850r,discplot850theta, nrow=2)

```


Figure \@ref(fig:discplot850) has smoother changes over time than in Figure \@ref(fig:discplot750), where $\delta$ = 0.750, which could indicate that there is less overfitting. One surprising result from this model is how the distance parameters slightly increase over time, and the intercepts slightly decrease. This could be a result of team shot selection evolving throughout the season, or just a coincidental signal.

## Conclusion

The evaluations of the models show that there is some weak evidence for time-dependency in shooting success rate in this dataset of player-tracking data from the Duke Men's Basketball team. Allowing predictors of shot success to shift based on recent success does not significantly improve the predictive accuracy of a model. However, we do see a systematic improvement in likelihood for smaller discount factors (i.e., more emphasis on recent shots, and therefore support of "streakiness"). Weaknesses of the discounting model include a smaller sample size and poorer out-of-sample prediction. Takeaways that we observed in other models include the fact that the angle of the shot only matters for certain players, and it is not a significant predictor of shot success between all players. Also, the effects of home-court advantage are not strong in this dataset, possibly due to the fact that most of the games away from home are missing.

To account for possible unexplained variation between seasons, and for variation introduced from having such a small population of road games, I repeated this analysis on a subset of the data that only consisted of shots from available games in one season (25 games), and shots from all home games (82 games). The results were similar, except for increased uncertainty due to smaller sample sizes. The model evaluation plots show similar patterns to the ones in Figure \@ref(fig:evalplot), and they are presented in Appendix B.

## Future Goals

Future goals for this research are to build a better-fitting model to predict basketball shots using more advanced factors that can be approximated from the dataset. Possibilities for this include reparametrizing the location of a shot using categories (e.g., corner three-point shots, heaves from half-court) using the distance of the nearest defender as a proxy for defense quality, or using the amount of time a player has played without a substitution or timeout to approximate fatigue. In addition, we make the unrealistic assumption that every player exhibits the same amount of shooting streakiness with our discounted likelihood models. We could improve upon this assumption in the discounted likelihood models by adding a random effect on $\delta$, to allow it to vary by player.
