==> index/06-appendix.Rmd <==
`r if(knitr:::is_latex_output()) '\\appendix'`

`r if(!knitr:::is_latex_output()) '# (APPENDIX) Appendix {-}'` 

# Code for Models

## Generalized Linear Model

```{r eval=FALSE, echo=TRUE}

Xtrainsub <- 
  Xtrain %>% 
  filter(as.integer(as.factor(gameid)) < 5)

priormod <- glm(formula = result ~ log(r) + theta, 
                data = Xtrainsub, 
                family = "binomial")

mu0r <- summary(priormod)[["coefficients"]]["log(r)","Estimate"]
mu0theta <- summary(priormod)[["coefficients"]]["theta","Estimate"]

fit_glm <- function(dat, S = 10000, B = 500){

  model.glm <- function(){
  
    # Liklihood function (for N observations)
    for(i in 1:N){
      
      logit(prob[i]) <- beta_int*int[i] + 
        beta_home*home[i] +
        beta_r*logr[i] + 
        beta_theta*theta[i]
      
      result[i] ~ dbern(prob[i])
    }

    # Priors
    # we expect less variation in the distance parameter, 
    # because shot success rate should get worse 
    # as distance increases under baseline circumstances.
    beta_int   ~ dnorm(0, 0.1)
    beta_home  ~ dnorm(0, 0.1)
    beta_r     ~ dnorm(mu0r, 0.01) 
    beta_theta ~ dnorm(mu0theta, 0.1)
  }

  datlist.glm <-  
    list(
      int = rep(1, nrow(dat)),
      logr = log(dat$r), 
      theta = dat$theta, 
      result = dat$result,
      home = dat$home,
      N = nrow(dat), 
      mu0r = mu0r,
      mu0theta = mu0theta
    )
  
  params.glm <- c("beta_int",
                  "beta_home",
                  "beta_r",
                  "beta_theta")

  initslist <- list(list("beta_int"=0, 
                         "beta_r"=0, 
                         "beta_theta"=0, 
                         "beta_home"=0))
                    
  sim <- 
    jags(data = datlist.glm, 
         n.chains = 1, n.iter = S, n.burnin = B, n.thin = 1,
         inits=initslist,
         parameters.to.save = params.glm,
         model.file=model.glm
    )
  
  sim.mcmc <- as.data.frame(as.mcmc(sim)[[1]])
  
  # Changing from a baseline mean + a shift amount
  # to two different means based on the type of game.
  sim.mcmc <- 
    sim.mcmc %>% 
    mutate(beta_intA = beta_int,
           beta_intH = beta_int + beta_home) %>%
    select(beta_intA, beta_intH, beta_r, beta_theta)
  
  return(sim.mcmc)
}

```

## Hierarchical Generalized Linear Model

```{r eval=FALSE, echo=TRUE}

Xtrainsub <- 
  Xtrain %>% 
  filter(as.integer(as.factor(gameid)) < 5)

priormod <- glm(formula = result ~ log(r) + theta, 
                data = Xtrainsub, 
                family = "binomial")

mu0r <- summary(priormod)[["coefficients"]]["log(r)","Estimate"]
mu0theta <- summary(priormod)[["coefficients"]]["theta","Estimate"]

fit_players <- function(dat = NA, S = 10000, B = 500){
  
  model.player <- function(){
    
    # Likelihood function for N observations
    for(i in 1:N){
      
      # the parameters now vary by player id.
      logit(prob[i]) <- beta_int[player[i]]*int[i] + 
        beta_home[player[i]]*home[i] + 
        beta_r[player[i]]*logr[i] + 
        beta_theta[player[i]]*theta[i]
      
      result[i] ~ dbern(prob[i])
    }
    
    # Priors
    for(j in 1:M){
        beta_int[j] ~ dnorm(beta_int0,tau_int)
        beta_home[j]  ~ dnorm(beta_home0, tau_int)
        beta_r[j] ~ dnorm(beta_r0,tau_r)
        beta_theta[j] ~ dnorm(beta_theta0,tau_theta)
    }
    
    # Hyperpriors
    beta_int0   ~ dnorm(0, 0.1)
    beta_home0  ~ dnorm(0, 0.1)
    beta_r0     ~ dnorm(mu0r, 0.01)
    beta_theta0 ~ dnorm(mu0theta, 0.1)
    tau_int ~ dgamma(10, 100)
    tau_r ~ dgamma(10, 0.2)
    tau_theta ~ dgamma(10, 10)
  }

  datlist.player <- 
    list(
      logr = log(dat$r), 
      theta = dat$theta, 
      home = dat$home,
      result = dat$result, 
      player = as.integer(as.factor(dat$globalplayerid)),
      N = nrow(dat), 
      int = rep(1, nrow(dat)), 
      M = n_distinct(dat$globalplayerid),
      mu0r = mu0r,
      mu0theta = mu0theta
    )
  
  # we want posteriors for the overall effects 
  # and for the individual player effects
  params <- c("beta_int", 
              "beta_home", 
              "beta_r", 
              "beta_theta",
              "beta_int0",
              "beta_home0",
              "beta_r0", 
              "beta_theta0")

  M <- datlist.player$M
  
  initslist <- list(
    list("beta_int"=rep(0,M), 
         "beta_home"=rep(0,M), 
         "beta_r"=rep(0,M), 
         "beta_theta"=rep(0,M),
         "beta_int0"=0,
         "beta_home0"=0,
         "beta_r0"=0, 
         "beta_theta0"=0, 
         "tau_int"=1, 
         "tau_r"=1, 
         "tau_theta"=1
  ))

  sim.player <- 
    jags(data = datlist.player, 
         n.iter = S, n.chains = 1, n.burnin = B, n.thin = 1,
         inits=initslist,
         parameters.to.save = params,
         model.file=model.player
  )
  
  sim.mcmc.player <- as.data.frame(as.mcmc(sim.player)[[1]])
  
  
  # Changing from a baseline mean + a shift amount
  # to two different means based on the type of game.
  hometext <- paste0("`beta_intH[",1:M,"]` = 
                     `beta_int[",1:M,"]` + 
                     `beta_home[",1:M,"]`",
                     collapse=",\n")
  
  awaytext <- paste0("`beta_intA[",1:M,"]` = 
                     `beta_int[",1:M,"]`", 
                     collapse=",\n")
  
  sim.mcmc.player <- eval(parse(text=
    paste0("sim.mcmc.player %>%
              mutate(",hometext,",
                     beta_intH0 = beta_int0 + beta_home0)","%>%
              rename(",awaytext,",
                     beta_intA0 = beta_int0)"
           ))) %>%
    select(grep("(beta_int)|(beta_theta)|(beta_r)",names(.)))
  
  colorder <- order(colnames(sim.mcmc.player))
  sim.mcmc.player <- sim.mcmc.player[ , colorder]
  
  
  # Renaming mixed effects columns from default factor levels 
  # (integers) to the corresponding player ids
  
  factorids <- 
    str_extract_all(names(sim.mcmc.player), "[[:digit:]]+") %>% 
    as.numeric()
  
  fids <- data.frame(factorid = factorids, 
                     order = 1:length(factorids))
  
  datmap <- dat %>% 
    mutate(factorid = as.integer(as.factor(globalplayerid))) %>%
    select(globalplayerid, factorid)
  
  gameids <- merge(datmap, fids, all.x=FALSE,all.y=TRUE) %>% 
    unique() %>% 
    mutate(globalplayerid = ifelse(is.na(globalplayerid),
                                   0,
                                   globalplayerid)) %>% 
    arrange(order)
  
  names(sim.mcmc.player) <- 
    str_replace_all(names(sim.mcmc.player), "[[:digit:]]+",
                    as.character(gameids$globalplayerid))
  
  return(sim.mcmc.player)
  
}


```

## Discounted Likelihood Hierarchical Model

```{r eval=FALSE, echo=TRUE}

fit_game <- function(dat = NA, g0 = NA, delta = NA, 
                     S = 10000, B = 500){

  model.game <- function(){

    for(i in 1:N){

      # delta = discount rate for game g relative to anchor game g0
      wt[i] <- delta^abs(games[i]-g0)  
      
      # player-level random effects
      logit(prob[i]) <- beta_int[player[i]]*int[i] + 
                        beta_home[player[i]]*home[i] + 
                        beta_r[player[i]]*logr[i] + 
                        beta_theta[player[i]]*theta[i] 

      # likelihood function
      p1[i] <- prob[i]^result[i]
      p2[i] <- (1-prob[i])^(1-result[i])
      
      # discounted likelihood function
      pi[i] <- (p1[i] * p2[i])^wt[i]  
      
      # defines correct discounted likelihood function
      y[i] ~ dbern(pi[i]) 

    }
    
    # Priors
    for(j in 1:M){
      beta_int[j] ~ dnorm(beta_int0,tau_int)
      beta_home[j] ~ dnorm(beta_home0, tau_int)
      beta_r[j] ~ dnorm(beta_r0,tau_r)
      beta_theta[j] ~ dnorm(beta_theta0,tau_theta)
    }
    
    # Hyperoriors
    beta_int0   ~ dnorm(0, 0.1)
    beta_home0  ~ dnorm(0, 0.1)
    beta_r0     ~ dnorm(mu0r, 0.01)
    beta_theta0 ~ dnorm(mu0theta, 0.1)
    tau_int ~ dgamma(10, 100)
    tau_r ~ dgamma(10, 0.2)
    tau_theta ~ dgamma(10, 10)
  }
  
  datlist.game <- 
    list(
      int = rep(1, nrow(dat)), 
      logr = log(dat$r), 
      theta = dat$theta, 
      result = dat$result, 
      home = dat$home,
      player = as.integer(as.factor(dat$globalplayerid)),
      N = nrow(dat), 
      M = n_distinct(dat$globalplayerid),
      mu0r = mu0r,
      mu0theta = mu0theta,
      delta = delta,
      games = as.integer(as.factor(dat$gameid)),
      g0 = g0,
      y = rep(1, nrow(dat))
    )

  params <- c("beta_int",
              "beta_r", 
              "beta_home",
              "beta_theta",
              "beta_int0",
              "beta_home0",
              "beta_r0", 
              "beta_theta0")

  M <- n_distinct(dat$globalplayerid)
  
  initslist <- list(list("beta_int"=rep(0,M), 
                         "beta_r"=rep(0,M), 
                         "beta_theta"=rep(0,M),
                         "beta_int0"=0,
                         "beta_r0"=0,
                         "beta_theta0"=0, 
                         "tau_int"=1, 
                         "tau_r"=1, 
                         "tau_theta"=1
                         ))

  sim.game <- jags(data = datlist.game, 
              n.iter = S, n.chains = 1, n.burnin = B, n.thin = 1,
              inits = initslist,
              parameters.to.save = params,
              model.file=model.game
  )
  sim.mcmc.game <- as.data.frame(as.mcmc(sim.game)[[1]])

  # Changing from a baseline mean + a shift amount
  # to two different means based on the type of game.
  
  hometext <- paste0("`beta_intH[",1:M,"]` = 
                     `beta_int[",1:M,"]` + 
                     `beta_home[",1:M,"]`", 
                     collapse=",\n")
  
  awaytext <- paste0("`beta_intA[",1:M,"]` = 
                     `beta_int[",1:M,"]`", 
                     collapse=",\n")

  sim.mcmc.game <- eval(parse(text=
    paste0("sim.mcmc.game %>%
              mutate(",hometext,",
                     beta_intH0 = beta_int0 + beta_home0)"," %>%
              rename(",awaytext,",
                     beta_intA0 = beta_int0)"
           ))) %>% 
    select(grep("(beta_int)|(beta_theta)|(beta_r)",names(.)))
  
  colorder <- order(colnames(sim.mcmc.game))
  sim.mcmc.game <- sim.mcmc.game[ , colorder]

  # Renaming mixed effects columns from default factor levels
  # (integers) to the corresponding player ids

  
  factorids <- 
    str_extract_all(names(sim.mcmc.game), "[[:digit:]]+") %>%
    as.numeric()
  fids <- data.frame(factorid = factorids, 
                     order = 1:length(factorids))
  
  datmap <- dat %>% 
    mutate(factorid = as.integer(as.factor(globalplayerid))) %>%
    select(globalplayerid, factorid)
  
  gameids <- merge(datmap, fids, all.x=FALSE,all.y=TRUE) %>% 
    unique() %>% 
    mutate(globalplayerid = ifelse(is.na(globalplayerid),
                                   0,
                                   globalplayerid)) %>% 
    arrange(order)
  
  names(sim.mcmc.game) <- 
    str_replace_all(names(sim.mcmc.game), 
                    "[[:digit:]]+",
                    as.character(gameids$globalplayerid))

  return(sim.mcmc.game)

}

```

```{r diagnostics, echo=FALSE, include=FALSE, cache=TRUE}

diagnostic_plots <- function(chain=NA, chain_name = ""){
  
  chain_df <- as.data.frame(chain)
  
  density <- ggplot(data=chain_df, aes(x=chain)) + geom_density() + 
    labs(title=paste("Posterior Density of", chain_name),
         x=chain_name, 
         y="Density") + 
    theme_bw() +
    theme(panel.grid = element_blank()) +
    geom_vline(xintercept=0, linetype=1, color="red") 

  trace <- ggplot(data=chain_df, aes(y=chain, x=1:length(chain))) + 
    geom_line() + 
    labs(title=paste("Trace Plot of", chain_name),
         x="Iteration", 
         y=chain_name) + 
    theme_bw() +
    theme(panel.grid = element_blank()) +
    geom_abline(intercept=0, slope=0, linetype=1, color="red") 
  lagautocorr <- ggplot(data=as.data.frame(1:50), aes(x=1:50, y=as.numeric(autocorr(as.mcmc(chain), lags = 1:50)))) + 
    geom_col(fill="black") + 
    labs(title=paste("Lag-t Autocorrelation of", chain_name),
         x="t", 
         y="Autocorrelation") + 
    theme_bw() +
    theme(panel.grid = element_blank()) +
    coord_cartesian(ylim = c(-1, 1)) +
    geom_abline(intercept=0, slope=0, linetype=1, color="red") 
  return(grid.arrange(density, trace, lagautocorr, nrow=3))
}

# load(file=paste0(rdatafiles,"glmtot.RData"))
# load(file=paste0(rdatafiles,"playermcmc.RData"))
# game.mcmc.list.750 <- load_gamelist(750)

# glmdiag1 <- diagnostic_plots(glmtot[[1]], "Intercept (Away)")
# glmdiag2 <- diagnostic_plots(glmtot[[2]], "Intercept (Home)")
# glmdiag3 <- diagnostic_plots(glmtot[[3]], "Distance")
# glmdiag4 <- diagnostic_plots(glmtot[[4]], "Angle")
# 
# # grid.arrange(glmdiag1, glmdiag2, glmdiag3, glmdiag4, nrow=2, top="GLM")
# 
# mediag1 <- diagnostic_plots(player.mcmc[paste0("beta_intA[",id1,"]")], "Intercept (Away)")
# mediag2 <- diagnostic_plots(player.mcmc[paste0("beta_intH[",id1,"]")], "Intercept (Home)")
# mediag3 <- diagnostic_plots(player.mcmc[paste0("beta_r[",id1,"]")], "Distance")
# mediag4 <- diagnostic_plots(player.mcmc[paste0("beta_theta[",id1,"]")], "Angle")

# grid.arrange(mediag1, mediag2, mediag3, mediag4, nrow=2, top="Mixed Effects (Player 1)")

# discdiag1 <- diagnostic_plots(game.mcmc.list.750[[10]][paste0("beta_intA[",id1,"]")], "Intercept (Away)")
# discdiag2 <- diagnostic_plots(player.mcmc[paste0("beta_intH[",id1,"]")], "Intercept (Home)")
# discdiag3 <- diagnostic_plots(player.mcmc[paste0("beta_r[",id1,"]")], "Distance")
# discdiag4 <- diagnostic_plots(player.mcmc[paste0("beta_theta[",id1,"]")], "Angle")

# grid.arrange(discdiag1, discdiag2, discdiag3, discdiag4, nrow=2, top="Discounted Weight Effects (Player 1)")

```

```{r diagplots, fig.height=5}

# grid.arrange(glmdiag1, glmdiag2, glmdiag3, glmdiag4, nrow=2, top="GLM")
# 
# grid.arrange(mediag1, mediag2, mediag3, mediag4, nrow=2, top="Mixed Effects (Player 1)")
# 
# grid.arrange(discdiag1, discdiag2, discdiag3, discdiag4, nrow=2, top="Discounted Weight Effects (Player 1, Game X)")


```

# Reproducing Evaluation Plots

## Home Games Only

```{r home, fig.height=6.5}

tempenv <- new.env()
eval.df <- NULL
  
for(i in 1:k){
    
  load(file = paste0(get_rdatafiles(i, home_only=TRUE), "evaldf.RData"), tempenv)
  eval.df <- rbind(eval.df, tempenv$eval.df)
    
}
    
eval.df$k <- as.factor(eval.df$k)
eval.df$delta <- factor(eval.df$delta,sort(as.character(levels(eval.df$delta)))) #making sure that the deltas are in order

df_delta <- eval.df %>% filter(!is.na(delta))
df <- eval.df %>% filter(is.na(delta))
gh1 <- ggplot(data=df_delta, aes(x=delta,y=class, col=k)) +
   geom_point() + 
   geom_hline(data=df, aes(yintercept=class, linetype=type, col=k)) +
   theme_bw() + 
   theme(axis.text.x = element_text(angle = 45, hjust=1)) +
   labs(y="Classification Rate \n (0.5 cutoff)", x=expression(delta), linetype = "")

gh2 <- ggplot(data=df_delta, aes(x=delta,y=lik, col=k)) +
  geom_point() +
  geom_hline(data=df, aes(yintercept=lik, linetype=type, col=k)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(y="Log Likelihood", x=expression(delta), linetype = "")

gh3 <- ggplot(data=df_delta, aes(x=delta,y=mse, col=k)) +
  geom_point() +
  geom_hline(data=df, aes(yintercept=mse, linetype=type, col=k)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(y="Brier Score (MSE)", x=expression(delta), linetype = "")

grid.arrange(gh1, gh2, gh3, nrow=2)

```

## One Season Only

```{r 2015, fig.height=6.5}

tempenv <- new.env()
eval.df <- NULL
  
for(i in 1:k){
    
  load(file = paste0(get_rdatafiles(i, season_2015=TRUE), "evaldf.RData"), tempenv)
  eval.df <- rbind(eval.df, tempenv$eval.df)
    
}
    
eval.df$k <- as.factor(eval.df$k)
eval.df$delta <- factor(eval.df$delta,sort(as.character(levels(eval.df$delta)))) #making sure that the deltas are in order

df_delta <- eval.df %>% filter(!is.na(delta))
df <- eval.df %>% filter(is.na(delta))
gy1 <- ggplot(data=df_delta, aes(x=delta,y=class, col=k)) +
   geom_point() + 
   geom_hline(data=df, aes(yintercept=class, linetype=type, col=k)) +
   theme_bw() + 
   theme(axis.text.x = element_text(angle = 45, hjust=1)) +
   labs(y="Classification Rate \n (0.5 cutoff)", x=expression(delta), linetype = "")

gy2 <- ggplot(data=df_delta, aes(x=delta,y=lik, col=k)) +
  geom_point() +
  geom_hline(data=df, aes(yintercept=lik, linetype=type, col=k)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(y="Log Likelihood", x=expression(delta), linetype = "")

gy3 <- ggplot(data=df_delta, aes(x=delta,y=mse, col=k)) +
  geom_point() +
  geom_hline(data=df, aes(yintercept=mse, linetype=type, col=k)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(y="Brier Score (MSE)", x=expression(delta), linetype = "")

grid.arrange(gy1, gy2, gy3, nrow=2)

```
